+++ backend/lib/migrate_template.js
@@ -56,4 +56,4 @@ const down = (_knex) => {
 	return Promise.resolve(true);
 };
 
-export { up, down };
+export { down, up };